% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_derivatives.R
\name{gaussian_loss_gradient}
\alias{gaussian_loss_gradient}
\title{Compute the gradient of the Gaussian loss function
with respect to the inverse covariance matrix}
\usage{
gaussian_loss_gradient(Sigma, X)
}
\arguments{
\item{Sigma}{A p x p covariance matrix}

\item{X}{A n x p data matrix}
}
\value{
A p x p matrix representing the gradient of the Gaussian loss function
}
\description{
Compute the gradient of the Gaussian loss function
with respect to the inverse covariance matrix
}
