% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_derivatives.R
\name{gaussian_loss_hessian}
\alias{gaussian_loss_hessian}
\title{Compute the Hessian of the Gaussian loss function
with respect to the inverse covariance matrix}
\usage{
gaussian_loss_hessian(Sigma)
}
\arguments{
\item{Sigma}{A p x p covariance matrix}
}
\value{
A (p^2) x (p^2) matrix representing the Hessian of the Gaussian loss function
}
\description{
Compute the Hessian of the Gaussian loss function
with respect to the inverse covariance matrix
}
