% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_graph.R
\name{graph_estimate}
\alias{graph_estimate}
\title{Graph estimation within the selected graph}
\usage{
graph_estimate(X, selected, sandwich.variance)
}
\arguments{
\item{X}{Data matrix (n x p)}

\item{selected}{Output from graphSelect function}

\item{sandwich.variance}{Logical indicating whether to use sandwich variance estimator (default: FALSE)}
}
\value{
A list containing the sample covariance matrix Sn, one-step
}
\description{
Estimate the precision matrix and variance within the selected graph structure.
}
