% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_Gaussian}
\alias{inference_Gaussian}
\title{Get Gaussian p-value and confidence interval}
\usage{
inference_Gaussian(theta, se, nullvalue, alpha)
}
\arguments{
\item{theta}{Estimated parameter}

\item{se}{Standard error of the estimated parameter}

\item{nullvalue}{Null hypothesis value for the parameter}

\item{alpha}{Significance level for confidence intervals}
}
\value{
A list containing the p-value, lower and upper bounds of the confidence interval
}
\description{
Get Gaussian p-value and confidence interval
}
