% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_truncatedGaussian}
\alias{inference_truncatedGaussian}
\title{Get truncated Gaussian p-value and confidence interval}
\usage{
inference_truncatedGaussian(theta, j, Var, nullvalue, A, b, alpha)
}
\arguments{
\item{theta}{Estimated parameter}

\item{j}{Index of the parameter in the full parameter vector}

\item{Var}{Covariance matrix of the estimated parameters}

\item{nullvalue}{Null hypothesis value for the parameter}

\item{A}{Affine constraint matrix}

\item{b}{Affine constraint vector}

\item{alpha}{Significance level for confidence intervals}
}
\value{
A list containing the p-value, lower and upper bounds of the confidence interval
}
\description{
Get truncated Gaussian p-value and confidence interval
}
